Updated post processing script
